This largely works
There is still an appdirs hack.  I don't know _why_ it is required.  Without
it, there is no appdirs package available.

The main cause of the problem was a nixpkgs bug. :(  I don't know _what_ the
bug was.  However, bumping nixpkgs to a recent master revision made the most
eggregious parts of the problem go away. :(